# Move regex opt to after lookup