@gouthamve 2020-08-04
```
➜  cortex git:(fix-2906) ✗ benchcmp old.txt new.txt 
benchmark                                     old ns/op     new ns/op     delta
BenchmarkParseIndexEntries500-8               342057        291739        -14.71%
BenchmarkParseIndexEntries2500-8              1718683       1516488       -11.76%
BenchmarkParseIndexEntries10000-8             6568014       6463280       -1.59%
BenchmarkParseIndexEntries50000-8             34212971      34284051      +0.21%
BenchmarkParseIndexEntriesRegexSet500-8       137045        89742         -34.52%
BenchmarkParseIndexEntriesRegexSet2500-8      647045        441737        -31.73%
BenchmarkParseIndexEntriesRegexSet10000-8     2772264       1881870       -32.12%
BenchmarkParseIndexEntriesRegexSet50000-8     14044719      10336336      -26.40%

benchmark                                     old allocs     new allocs     delta
BenchmarkParseIndexEntries500-8               1503           1504           +0.07%
BenchmarkParseIndexEntries2500-8              7503           7504           +0.01%
BenchmarkParseIndexEntries10000-8             30005          30007          +0.01%
BenchmarkParseIndexEntries50000-8             150007         150008         +0.00%
BenchmarkParseIndexEntriesRegexSet500-8       1503           1522           +1.26%
BenchmarkParseIndexEntriesRegexSet2500-8      7503           7522           +0.25%
BenchmarkParseIndexEntriesRegexSet10000-8     30005          30023          +0.06%
BenchmarkParseIndexEntriesRegexSet50000-8     150007         150024         +0.01%

benchmark                                     old bytes     new bytes     delta
BenchmarkParseIndexEntries500-8               96289         96320         +0.03%
BenchmarkParseIndexEntries2500-8              481975        482008        +0.01%
BenchmarkParseIndexEntries10000-8             1928507       1928559       +0.00%
BenchmarkParseIndexEntries50000-8             9606750       9606808       +0.00%
BenchmarkParseIndexEntriesRegexSet500-8       88222         88693         +0.53%
BenchmarkParseIndexEntriesRegexSet2500-8      441163        441627        +0.11%
BenchmarkParseIndexEntriesRegexSet10000-8     1764739       1765087       +0.02%
BenchmarkParseIndexEntriesRegexSet50000-8     8804033       8804335       +0.00%
```

Hrm, not such a big change, interesting but still worth having imo.