@pstibrany 2020-08-20
This looks useful, and would be great to merge. @gouthamve would you find some time to look at the comments and get the PR into mergeable state?